add place order api

+++ src/main/java/com/example/Swigato/model/OrderEntity.java
@@ -24,7 +24,7 @@ public class OrderEntity {
 
     String orderId;      //UUID
 
-    int orderTotal;
+    double orderTotal;
 
     @CreationTimestamp
     Date orderTime;
